Acréscimos de relatórios Pivot e finalização das views
Criação de relatórios e colocação de views mais estilizadas.

+++ Models/Relatorio.cs
@@ -1,8 +1,8 @@
 ﻿namespace Projeto22025.Models
 {
-    public class Relatorio
+     public class Relatorio
     {
         public string Grupo { get; set; } = string.Empty;
-        public int Total { get; set; }
+        public double Total { get; set; }
     }
 }
